requirements.txt
python-telegram-bot==21.9
httpx==0.28.1
tenacity==9.0.0
python-dotenv==1.0.1
